# backend/api/src/main/resources/openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GPT Lambda API",
    "description": "GPT Lambda AI API Service",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.functionhub.net",
      "description": "Prod"
    },
    {
      "url": "http://localhost:8080",
      "description": "Dev"
    }
  ],
  "paths": {
...
    "/status": {
      "post": {
        "tags": [
          "spec"
        ],
        "operationId": "getSpecStatus",
        "description": "Get user spec generation status",
        "requestBody": {
          "required": true,
          "description": "Status request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          }
        }
      }
    },
    "/{function_slug}": {
      "post": {
        "tags": [
          "runtime"
        ],
        "summary": "Execute any user function",
        "description": "Execute any user function with dynamically generated request and response payloads",
        "operationId": "runProdFunction",
        "parameters": [
          {
            "name": "function_slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Execution request",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/d/{function_slug}": {
      "post": {
        "tags": [
          "runtime"
        ],
        "summary": "Execute any user function",
        "description": "Execute any user function with dynamically generated request and response payloads",
        "operationId": "runDevFunction",
        "parameters": [
          {
            "name": "function_slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Execution request",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/spec/{function_slug}/{version}/{auth_token}/{env}": {
      "get": {
        "tags": [
          "spec"
        ],
        "operationId": "getUserSpec",
        "parameters": [
          {
            "name": "function_slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "auth_token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "env",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User code",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/npm/{uid}": {
      "get": {
        "tags": [
          "runtime"
        ],
        "operationId": "getUserCode",
        "parameters": [
          {
            "name": "uid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User code",
            "content": {
              "application/typescript": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/e-result": {
      "get": {
        "tags": [
          "runtime"
        ],
        "operationId": "getExecResult",
        "summary": "Get result of completed execution",
        "description": "Execution results are short-lived",
        "parameters": [
          {
            "name": "exec_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecResultAsync"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "runtime"
        ],
        "summary": "Execution result handler",
        "description": "Handles user code execution result",
        "operationId": "handleExecResult",
        "requestBody": {
          "required": true,
          "description": "Execution result",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecResultAsync"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Runtime Info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
    "/s-result": {
      "post": {
        "tags": [
          "runtime"
        ],
        "summary": "Spec result handler",
        "description": "Handles spec generation result",
        "operationId": "handleSpecResult",
        "requestBody": {
          "required": true,
          "description": "Spec result",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpecResult"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Runtime Info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericResponse"
                }
              }
            }
          }
        }
      }
    },
    "/update-code": {
      "post": {
        "tags": [
          "runtime"
        ],
        "summary": "Update code",
        "description": "Update code",
        "operationId": "updateCode",
        "requestBody": {
          "required": true,
          "description": "Update request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Code"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Code update response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeUpdateResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/code/{uid}": {
      "get": {
        "tags": [
          "runtime"
        ],
        "summary": "Get user code detail",
        "description": "Get user code detail",
        "operationId": "getCodeDetail",
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Code update response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Code"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Code":  {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string",
            "description": "Code cell uid"
          },
          "code": {
            "type": "string"
          },
          "is_public": {
            "type": "boolean"
          },
          "is_active": {
            "type": "boolean"
          },
          "updated_at": {
            "type": "integer",
            "format": "int64"
          },
          "created_at": {
            "type": "integer",
            "format": "int64"
          },
          "parent_id": {
            "type": "string"
          },
          "function_slug": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "fields_to_update": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "StatusResponse": {
        "type": "object",
        "properties": {
          "is_ready": {
            "type": "boolean"
          },
          "error": {
            "type": "string"
          }
        }
      },
      "StatusRequest": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "deployed": {
            "type": "boolean"
          }
        }
      },
      "CodeUpdateResult": {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "project_id": {
            "type": "string"
          }
        }
      },
      "ExecResultAsync":  {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string"
          },
          "exec_id": {
            "type": "string"
          },
          "validate": {
            "type": "boolean",
            "default": false
          },
          "result": {
            "type": "string"
          },
          "std_out": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "std_out_str": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        }
      },
      "Spec":  {
        "type": "object",
        "properties": {
          "value": {
            "type": "string"
          },
          "format": {
            "type": "string"
          }
        }
      },
      "SpecResult":  {
        "type": "object",
        "properties": {
          "uid": {
            "type": "string"
          },
          "spec": {
            "$ref": "#/components/schemas/Spec"
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "GenericResponse":  {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        }
      },
...
    }
  }
}